% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_n.R
\name{last_n}
\alias{last_n}
\title{Last n letters}
\usage{
last_n(s, n)
}
\arguments{
\item{s}{character vector.}

\item{n}{integer.}
}
\value{
character vector. The elements of the result vector are last n letters
 of input vector s.
}
\description{
Last n letters
}
\examples{
last_n(c("abc", "wxyz"), 2)
}
