% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching_compute}
\alias{matching_compute}
\title{マッチングを計算する}
\usage{
matching_compute(util, slots)
}
\arguments{
\item{util}{list. \code{\link[=matching_utils]{matching_utils()}} の結果}

\item{slots}{integer vector. 各ゼミの定員}
}
\value{
マッチングの結果。列 (\code{Student}, \code{Seminar}) を持つ tibble.
}
\description{
\code{matchingR::galeShapley.collegeAdmissions()} を使って学生最適なマッチングを計算する。
}
