% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching_utils}
\alias{matching_utils}
\title{効用を計算する}
\usage{
matching_utils(
  student_list,
  faculty_list,
  dir_student,
  dir_faculty,
  nc = 8,
  seed = NULL
)
}
\arguments{
\item{student_list}{path string. Admin-managed list of students and their GPA.}

\item{faculty_list}{path string. Admin-managed list of faculty.}

\item{dir_student}{path string. Directory for student preferences.}

\item{dir_faculty}{path string. Directory for faculty preferences.}

\item{nc}{integer. Length of registration ID.}

\item{seed}{integer. Used as a random seed.}
}
\value{
Returns matching results.
}
\description{
効用を計算する
}
